requirements.txt
numpy==1.26.4
plotly==5.24.1
scipy==1.14.1
streamlit==1.40.0
pillow==10.4.0
scikit-image==0.24.0
nibabel==5.3.2
opencv-python==4.10.0.84
albumentations==1.3.1
pycirclize==1.7.1
psutil==6.1.0
tqdm==4.67.0
sympy==1.13.1
pandas==2.2.3
kaleido==0.2.1
ray==2.38.0
ray[tune]
